py/requirements.txt
psd-tools>=1.9.24
Pillow>=10.0.0 